[15.0][MIG]code improvement hr_applicant
