+++ hr_applicant/LICENSE/LICENSE
@@ -1,5 +1,5 @@
     OpenERP, Open Source Management Solution
-    Copyright (C) 2020-Today Serpent Consulting Services Pvt. Ltd.
+    Copyright (C) 2022-Today Serpent Consulting Services Pvt. Ltd.
     (<http://www.serpentcs.com>)
 
     This program is free software: you can redistribute it and/or modify
